Alpha

+++ backend/network/Client.ts
@@ -5,9 +5,10 @@ import {Server} from "../Server";
 import {Handshake} from "../packets/Handshake";
 import NanoTimer from "nanotimer";
 import {EntityType} from "../enums/EntityType";
-import {Box} from "../entities/Box";
+import {Crate} from "../entities/Crate";
 import {ActionType} from "../enums/ActionType";
 import {ServerPackets} from "../enums/packets/ServerPackets";
+import {ClientPackets} from "../enums/packets/ClientPackets";
 
 export class Client {
     public socket: WebSocket<any>;
@@ -51,41 +52,49 @@ export class Client {
                     return;
                 }
                 const handshake = new Handshake([PACKET_TYPE, ...PACKET_DATA], this);
+                const player = this.server.players.find(player => player.data.token === handshake.token);
+
+                if(player) {
+                    if(player.client.isActive) {
+                        player.client.sendBinary(new Uint8Array([ClientPackets.STEAL_TOKEN]));
+                        if(player.client.isActive) player.client.socket.close();
+                    }
+                    this.player = player;
+                    player.client = this;
+                    handshake.response(this.player, true);
+                    return;
+                }
+                if(this.server.players.length >= 99) {
+                    return this.sendBinary(new Uint8Array([ClientPackets.FULL]))
+                }
+
                 this.player = new Player(this);
+
                 handshake.setupPlayer(this.player);
 
                 this.server.players.push(this.player);
                 this.server.entities.push(this.player);
 
                 handshake.response(this.player);
+                handshake.broadcastCosmetics(this.player);
+
             }
 
             this.receivePacket(PACKET_TYPE, PACKET, PACKET_DATA);
-        } catch (error) {
-            // Обработка ошибки
-        }
+        } catch (error) {}
     }
 
     public receivePacket(PACKET_TYPE: number, PACKET: any, PACKET_DATA: number[]) {
         this.packetsQty[PACKET_TYPE]++;
         this.packetsSize[PACKET_TYPE] += PACKET_DATA.length;
 
-        if (this.packetsQty[0] > 5) {
-            this.socket.close();
-            return;
-        }
-        if (this.packetsQty[3] > 10) {
-            this.socket.close();
-            return;
-        }
-        if (this.packetsQty[PACKET_TYPE] > 30) {
-            this.socket.close();
-            return;
-        }
+        // if (this.packetsQty[0] > 5) return this.socket.close();
+        // if (this.packetsQty[3] > 10) return this.socket.close();
+        // if (this.packetsQty[PACKET_TYPE] > 30) return this.socket.close();
 
         switch (PACKET_TYPE) {
             case ServerPackets.CHAT:
-                this.broadcast(JSON.stringify([0, this.player.id, PACKET]));
+                this.server.broadcast(JSON.stringify([0, this.player.id, PACKET]), false, this.socket);
                 break;
             case ServerPackets.MOVEMENT:
                 this.player.direction = PACKET;
@@ -94,67 +103,80 @@ export class Client {
                 this.player.angle = Number(PACKET) % 255;
                 break;
             case ServerPackets.ATTACK:
+                if(this.player.isCrafting) break;
+                this.player.attack = true;
                 this.player.angle = Number(PACKET) % 255;
                 this.player.action |= ActionType.ATTACK;
-                this.player.attackManager.isAttack = true;
+                this.server.combatSystem.handleAttack(this.player);
                 break;
             case ServerPackets.INTERACTION:
-                this.player.interactionManager.useItem(PACKET);
+                if(this.player.isCrafting) break;
+                this.server.interactionSystem.request(this.player, PACKET);
                 break;
-            case 6:
+            case ServerPackets.DROP_ONE_ITEM:
+                if(this.player.isCrafting) break;
                 if (this.player.inventory.items.has(PACKET))
-                    new Box(this.server, EntityType.CRATE, {
+                    new Crate(this.server, {
                         owner: this.player,
                         item: PACKET,
                         count: this.player.inventory.items.get(PACKET)
                     });
                 this.sendBinary(this.player.inventory.deleteItem(PACKET));
                 break;
-            case 14:
-                this.player.attackManager.isAttack = false;
+            case ServerPackets.CRAFT:
+                if(this.player.isCrafting) break;
+                this.server.craftSystem.handleCraft(this.player, PACKET);
                 break;
-            case 28:
+            case ServerPackets.RECYCLE_START:
+                if(this.player.isCrafting) break;
+                this.server.craftSystem.handleRecycle(this.player, PACKET);
+                break;
+            case ServerPackets.BUILD:
+                if(this.player.isCrafting) break;
+                this.server.buildingSystem.request(this.player, PACKET_DATA);
+                break;
+            case ServerPackets.STOP_ATTACK:
+                this.player.attack = false;
+                break;
+            case ServerPackets.DROP_ITEM:
+                if(this.player.isCrafting) break;
                 if (this.player.inventory.items.has(PACKET))
-                    new Box(this.server, EntityType.CRATE, {
+                    new Crate(this.server, {
                         owner: this.player,
                         item: PACKET,
                         count: 1
                     });
                 this.sendBinary(this.player.inventory.removeItem(PACKET, 1));
                 break;
-            case 36:
-                this.player.commandManager.handleCommand(PACKET);
+            case ServerPackets.CANCEL_CRAFT:
+                this.server.craftSystem.stopCraft(this.player);
+                break;
+            case ServerPackets.CONSOLE:
                 break;
         }
     }
 
     public onClose() {
         this.isActive = false;
-
-        this.player.action = 1;
-        new NanoTimer().setTimeout(
-            () => {
-                this.server.playerPool.deleteId(this.player.id);
-                this.server.players = this.server.players.filter((player) => player !== this.player);
-                this.server.entities = this.server.entities.filter((player) => player !== this.player);
-            },
-            [],
-            "0.1s"
-        );
     }
 
     public sendJSON(message: any) {
-        this.socket.send(JSON.stringify(message));
+        if (this.isActive && message) this.socket.send(JSON.stringify(message));
+    }
+
+    public sendU8(message: any){
+        if (this.isActive && message) this.socket.send(new Uint8Array(message), true);
+    }
+
+    public sendU16(message: any){
+        if (this.isActive && message) this.socket.send(new Uint16Array(message), true);
+    }
+
+    public sendU32(message: any){
+        if (this.isActive && message) this.socket.send(new Uint32Array(message), true);
     }
 
     public sendBinary(message: Uint8Array | Uint16Array | Uint32Array | undefined) {
         if (this.isActive && message) this.socket.send(message, true);
-    }
-
-    public broadcast(message: any) {
-        const clients = Array.from(this.server.wss.clients.values());
-        for (const client of clients) {
-            if (client.socket !== this.socket) client.socket.send(message);
-        }
     }
 }
